requirement.txt
Flask==3.1.2
mysql-connector-python==9.5.0
click==8.3.0
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3
werkzeug==3.1.3
blinker==1.9.0
colorama==0.4.6
